requirements.txt
google-play-scraper
pandas
numpy
sentence-transformers
faiss-cpu
tqdm
openai
python-dotenv
